return OSM raster tiles for a given bounding box

return OSM raster tiles for a given bounding box

osm_tile(
  bbox = c(xmin = 4.26, ymin = 51.22, xmax = 4.5, ymax = 51.35),
  extra_zoom = 0,
  cache_path = R_user_dir("birdradar", which = "cache")
)

Arguments

bbox

Either (i) four numeric values specifying the maximal and minimal longitudes and latitudes, in the form c(xmin, ymin, xmax, ymax) or (ii) a character string in the form xmin,ymin,xmax,ymax. These will be passed to getbb to be converted to a numerical bounding box. Can also be (iii) a matrix representing a bounding polygon as returned from getbb(..., format_out = "polygon"). To search in an area, (iv) a character string with a relation or a (closed) way id in the format "way(id:1)", "relation(id:1, 2)" or "relation(id:1, 2, 3); way(id:2)" as returned by getbb(..., format_out = "osm_type_id") or bbox_to_string with a data.frame from getbb(..., format_out = "data.frame") to select all areas combined (relations and ways).

extra_zoom

the number of extra zoom levels

cache_path

File location to cache maps